refactor for default model config and model config get function for custom model settings
